move db specific utils to their own utilities file
